# Optimize levenshtein a bit for memory usage