--- a/ext/standard/levenshtein.c
+++ b/ext/standard/levenshtein.c
@@ -32,6 +32,15 @@ static zend_long reference_levdist(const zend_string *string1, const zend_string
 		return ZSTR_LEN(string1) * cost_del;
 	}
 
+	/* When all costs are equal, levenshtein fulfills the requirements of a metric, which means
+	 * that the distance is symmetric. If string1 is shorter than string 2 we can save memory (and CPU time)
+	 * by having shorter rows (p1 & p2). */
+	if (ZSTR_LEN(string1) < ZSTR_LEN(string2) && cost_ins == cost_rep && cost_rep == cost_del) {
+		const zend_string *tmp = string1;
+		string1 = string2;
+		string2 = tmp;
+	}
+
 	p1 = safe_emalloc((ZSTR_LEN(string2) + 1), sizeof(zend_long), 0);
 	p2 = safe_emalloc((ZSTR_LEN(string2) + 1), sizeof(zend_long), 0);